Combine all of the `coroutines` package into one file.
